feat: complete teams crud

+++ front/src/server/api/routers/teams.ts
@@ -18,22 +18,64 @@ export const teamsRouter = createTRPCRouter({
         },
       });
     }),
+  getTeamById: protectedProcedure
+    .input(z.object({ id: z.string().nullable().optional() }))
+    .query(({ input, ctx }) => {
+      if (!input.id) return null;
+      return ctx.prisma.team.findUnique({
+        where: {
+          id: input.id,
+        },
+        include: {
+          users: true,
+        },
+      });
+    }),
   createTeam: protectedProcedure
-    .input(z.object({ adminId: z.string(), name: z.string() }))
+    .input(
+      z.object({
+        adminId: z.string(),
+        name: z.string(),
+        users: z.string().array(),
+      })
+    )
     .mutation(({ input, ctx }) => {
       return ctx.prisma.team.create({
         data: {
           ...input,
+          users: {
+            connect: input.users.map((user) => ({ id: user })),
+          },
         },
       });
     }),
   deleteTeam: protectedProcedure
-    .input(z.object({ userId: z.string(), id: z.string() }))
+    .input(z.object({ userId: z.string().optional(), id: z.string() }))
     .mutation(({ input, ctx }) => {
+      if (!input.userId) return;
       return ctx.prisma.team.delete({
         where: {
           adminId: input.userId,
           id: input.id,
+        },
+      });
+    }),
+  updateTeam: protectedProcedure
+    .input(
+      z.object({
+        id: z.string(),
+        name: z.string(),
+        users: z.string().array(),
+      })
+    )
+    .mutation(({ input, ctx }) => {
+      return ctx.prisma.team.update({
+        where: {
+          id: input.id,
+        },
+        data: {
+          ...input,
+          users: { connect: input.users.map((user) => ({ id: user })) },
         },
       });
     }),
